src/posts/pyproject.toml
[project]
name = "posts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "faker>=38.2.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "protobuf>=6.33.2",
]
